Switched data type from double to float

+++ NeuralNetwork.NET/Convolution/Operations/KernelConvolution.cs
@@ -12,14 +12,14 @@ namespace NeuralNetworkNET.Convolution.Operations
         /// Gets the kernels chosen for the current instance
         /// </summary>
         [NotNull]
-        public double[][,] Kernels { get; }
+        public float[][,] Kernels { get; }
 
         /// <summary>
         /// Creates a new instance with the given kernels and type
         /// </summary>
         /// <param name="type">The type of operation</param>
         /// <param name="kernels">The kernels to use</param>
-        internal KernelConvolution(ConvolutionOperationType type, [NotNull] double[][,] kernels) : base(type)
+        internal KernelConvolution(ConvolutionOperationType type, [NotNull] float[][,] kernels) : base(type)
         {
             if (type != ConvolutionOperationType.Convolution3x3) throw new ArgumentOutOfRangeException("Invalid operation type");
             Kernels = kernels;
