+++ NeuralNetwork.NET/Networks/Activations/ActivationFunction.cs
@@ -4,5 +4,5 @@
     /// A delegate that represents an activation function (or the derivative of an activation function) used in a neural network
     /// </summary>
     /// <param name="x">The input value</param>
-    public delegate double ActivationFunction(double x);
+    public delegate float ActivationFunction(float x);
 }
